Refactor Symphony handlers into modular domain-organized structure
Split the monolithic `symphony.ts` (3315 lines) into six focused domain modules plus a shared utilities file and thin composer:

- **symphony/index.ts**: Thin composer that registers all domains
- **symphony/shared.ts**: Shared utilities (paths, state I/O, git ops, auth checks)
- **symphony/discovery.ts**: Registry/issue browsing (getRegistry, getIssues, getIssueCounts)
- **symphony/dashboard.ts**: Dashboard state (getState, getActive, getCompleted, getStats)
- **symphony/lifecycle.ts**: Contribution lifecycle transitions (registerActive, updateStatus, complete, cancel)
- **symphony/contributionStart.ts**: Contribution kickoff (start, cloneRepo, startContribution)
- **symphony/contributionFinish.ts**: Contribution completion (createDraftPR, fetchDocumentContent, manualCredit)
- **symphony/sync.ts**: PR syncing (checkPRStatuses, syncContribution, clearCache)

Each module exports a single `register*Handlers()` function, called from the composer. Shared helpers (state I/O, git operations, GitHub auth, path resolution) moved to `shared.ts` and are imported by multiple domains. No behavioral changes; purely organizational.
